requirements.txt
numpy==1.24.2
open3d
ConfigArgParse==1.5.3
image
Pillow==9.4.0